Experimental: split contract into execute, query, contract_tests
