+++ contracts/cw721-base/src/lib.rs
@@ -1,9 +1,13 @@
-pub mod contract;
+mod contract_tests;
 mod error;
+mod execute;
 pub mod msg;
+mod query;
 pub mod state;
 
 pub use crate::error::ContractError;
+pub use crate::msg::{ExecuteMsg, InstantiateMsg, MintMsg, MinterResponse, QueryMsg};
+pub use crate::state::Cw721Contract;
 
 pub mod entry {
     use super::*;
